refactor: testing out new derivation objects
